// src/types/generated/SessionSupervision.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SessionSupervision = { id: number, caissier_id: number, caissier_nom: string | null, magasin_nom: string | null, statut: string, date_ouverture: string, date_cloture: string | null, fond_initial: number, recettes_especes: number, recettes_cb: number, recettes_cheque: number, recettes_virement: number, sorties: number, entrees: number, especes_attendu: number | null, especes_declare: number | null, ecart: number | null, };